Add soln for 2022192
This is dreadfully slow. We're exploring a tree, using BFS and doing
some pruning:
- always build a geo bot if we can (Although this isn't optimal in
  general:
  https://www.reddit.com/r/adventofcode/comments/zpihwi/comment/j289m2b/)
- if we did nothing previous round but could have built a type of bot,
  don't build that type of bot in this round, since we should be greedy
- as we approach the end, only build bots that can increase the geo
  score

To improve this, we'd need to prune more aggresively. One way to do that
would be to change this to DFS, add a heuristic score to each candidate
(eg the maximum possible score by building geos for each of the remaining
rounds), explore the routes with the highest heuristics first to get a
best score, and then prune branches that can't reach that max score.